requirements.txt
PyJWT==2.7.0